:truck::recycle: Rename the Radio folder to match the component name
